strands-agents/probe42-strands-agent/pyproject.toml
[project]
name = "probe-strands-agent"
version = "1.0.0"
description = "Strands Agent for Probe42 API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.0.0",
    "httpx>=0.25.0",
    "streamlit>=1.28.0",
]

[project.license]
file = "LICENSE"

[project.scripts]
probe-agent = "probe_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/probe_agent"]